Bug fix: popup menu and blinking timer.
Bug reveal:
1. Start the program.
2. Click the main status item, and start a timer.
3. Before the timer finishes, click it to pop up its menu.
4. Keep the menu on until the timer finishes.
5. Click anywhere to cancel the popup menu.

Expected result:
The timer should be removed, since it's already finished, and the user acted to the blinking (by clicking to cancel the popup menu).

Actual result:
The timer leaves there, unhighlighted, with text "00:00". No way to remove the timer.

Cause:
By clicking to cancel the popup menu, no "Click-on-Status-Item" event is triggered. Only a popup-menu-disappeared event is triggered.

Fix:
Add "actionOnCancelPopped" action to the status item.